# perf: improve the performance of sending data