@lesismal 2024-05-23
OnData usually handles parsing, when we got a frame/message in OnData, we handle the frame/message async in another goroutine, which may not finish writing when OnData for next fd in the eventloop.

when:
1. len(writeList) == 0: this fast flush doesn't cost much;
2. len(wirteList) > 0: because the new writing maybe not finished as I explained above, so this may be the old data from old writing. at the beginning of this time of loop, there has check the write event, since still data waiting to be sent, that means this fd is still not writable. Then if we flush we'll waste more syscall.

The benchmark results are not different too much, and can not prove the promotion enough.
One more thing is, in the real scenarios over the public internet, TCP windows size communication is slower than in the local test env.