@limpo1989 2024-05-23
In LT mode, if there is no writeable event at present, it does not mean that the socket cannot be written, because the write event registered this time actually needs to wait until the next round of epoll-wait to be triggered.

Perhaps this change will improve intranet services, such as the rpc framework serving intranets. 🤣