@lesismal 2024-05-23
> In LT mode, if there is no writeable event at present, it does not mean that the socket cannot be written

No matter which epoll mod, It may be writable at the next nano when the epoll_wait returns and before we check the event type, but for most of the time, if there's no write event, it means not writable at this moment.
If we don't do flush according to the write event, then we don't need to check the write event and the kernal doesn't need to provide the write event anymore, because the fd maybe writable at any moment. :joy: